put tests in different file
